Correction algorithm based on merging chunks with good consensus.
New algorithm tries to preserve read length (% corrected errors).
Chunks are composed of subsequent 'good' hkmer cluster centers,
and aligned later against the read. If they don't overlap,
the part of the read between their respective ends is copied as is;
the case of overlap is not yet finished (scores are to be taken
into account) - this is one of main error sources, the other one
is absence of subclustering.